Data layer fixes

+++ app/src/main/java/com/kuznetsova/mlanning/data/TasksDataSource.kt
@@ -7,13 +7,13 @@ import com.kuznetsova.mlanning.domain.taskitem.TaskItem
 interface TasksDataSource {
 
     suspend fun getAllTask(): List<Task>
-    suspend fun getTaskById(id: Int): Task
+    suspend fun getTaskById(id: Long): Task
     suspend fun insertTask(task: Task)
-    suspend fun updateTask(taskId: Int, name: String, description: String?, priority: TaskPriority)
-    suspend fun deleteTask(id: Int)
+    suspend fun updateTask(taskId: Long, name: String, description: String?, priority: TaskPriority)
+    suspend fun deleteTask(id: Long)
 
-    suspend fun updateTaskItem(taskItem: TaskItem, taskId: Int)
-    suspend fun createTaskItem(taskItem: TaskItem, taskId: Int)
+    suspend fun updateTaskItem(taskItem: TaskItem, taskId: Long)
+    suspend fun createTaskItem(taskItem: TaskItem, taskId: Long)
     suspend fun deleteTaskItem(taskItemId: Int)
 
 }
